EHUB-178: issue fixed for delete and update transaction
